# Conflicts:
#	megano/auth_app/views/email_verify.py
#	megano/templates/profile_app/base.html

+++ megano/auth_app/repositories/auth_repositories.py
@@ -3,6 +3,8 @@ from auth_app.interfaces.auth_interface import IAuth
 from auth_app.models.user import User
 from django.core.exceptions import ObjectDoesNotExist
 from typing import Optional
+from datetime import date, timedelta
+from ..utils import get_activation_key
 
 
 class AuthRepository(IAuth):
@@ -20,14 +22,17 @@ class AuthRepository(IAuth):
             return None
 
     @beartype
-    def delete_user_by_email(self, _email: str) -> Optional[bool]:
+    def get_user(self, model: User) -> User:
+        """Получаем пользователя"""
+        user = model
+        return user
+
+    @beartype
+    def delete_user_by_email(self, _email: str) -> None:
         """Удаляем пользователя"""
-        try:
-            user = self.get_user_by_email(_email)
+        user = self.get_user_by_email(_email)
+        if user:
             user.delete()
-            return True
-        except ObjectDoesNotExist:
-            return None
 
     @beartype
     def get_user_by_activation_key(self, _activation_key: str) -> Optional[User]:
@@ -41,3 +46,16 @@ class AuthRepository(IAuth):
     def set_user_is_active(self, model: User, value: bool) -> None:
         """Установить значение параметра is_active пользователя"""
         model.is_active = value
+        self.save(model)
+
+    @beartype
+    def set_activation_key(self, model: User):
+        """Установить значение параметра activation_key пользователя"""
+        model.activation_key = get_activation_key(model.email)
+        self.save(model)
+
+    @beartype
+    def set_activation_key_will_expires(self, model: User):
+        """Установить значение параметра activation_key_will_expires пользователя"""
+        model.activation_key_will_expires = date.today() + timedelta(2)
+        self.save(model)
